Plaid: transitions file is now hardcoded
I wanted to use a GUI window to choose the transitions file. However, Tkinter (a native GUI library in Python) does not work properly in mac osx (within the Plaid script*), hence, for now, the transitions file will be hard coded. That means that the file 'transitions_file.txt' will be used, so you'll have to change the name of the transitions file that you intend to use to 'transitions_file.txt.\n\n* Note: Tkinter works by itself in a mac osx, but an error appears when Plaid runs with this gui.