libutils: DataStruct() does read parameters
Parameters used in the experiment will be written in the data file, but they are not being read in the analysis. That is because i do not know how to use np.genfromtxt() to read columns that contain more an array (e.g. color parameters). Hirherto experiment parameters have not been used in the analysis, so it is not a problem at all.