Plaid: fix reading parameters
When reading parameters from config and trials file, the length of the parameter's value was used to differenciate if the parameter was a single number or an array (e.g. colors). Now, for color parameters, the name of the parameter is used, check if 'color' is in the name. For the trials file, check if there are commas ',' in the parameter's value (e.g. lambda = 100,100,100 will be splited given the commas).