libutils: fix create_transitions file bug
When there was only one percept in some trial, there was an error message (too many indices) in the part of writing the durations of A and B to output array in create_transitions_file() and the transitions file was not created. Two additional if-else statements have been added to look out for the cases where A or B are empty.